Issue an error message and quit (rather than overflowing a reference counter) if the number of references to a table exceeds the maximum due to nested UNION views.  Fix for ticket [d58ccbb3f1].
FossilOrigin-Name: c2462a95ed8e1e69886681400d673207d906bf1b

+++ src/vdbeaux.c
@@ -1538,7 +1538,7 @@ void sqlite3VdbeMakeReady(
     zEnd = &zCsr[nByte];
   }while( nByte && !db->mallocFailed );
 
-  p->nCursor = (u16)nCursor;
+  p->nCursor = nCursor;
   p->nOnceFlag = nOnce;
   if( p->aVar ){
     p->nVar = (ynVar)nVar;
